restructured the dto package to seperate user related dtos from staff, manager and future needed dtos
