+++ src/main/java/com/oussama/space_renting/dto/AuthResponse.java
@@ -1,16 +1,17 @@
 package com.oussama.space_renting.dto;
 
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.stereotype.Service;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@Builder
 public class AuthResponse {
     private String message;
     private String token;
-
-    public AuthResponse(String message, String token) {
-        this.message = message;
-        this.token = token;
-    }
-
-    public String getMessage() { return message; }
-    public void setMessage(String message) { this.message = message; }
-    public String getToken() { return token; }
-    public void setToken(String token) { this.token = token; }
 }
